fixes + add brand crud

+++ src/Http/Controllers/Api/V1/BrandController.php
@@ -6,53 +6,60 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\JsonResource;
 use Illuminate\Validation\Rule;
+use PictaStudio\VenditioCore\Dto\BrandDto;
 use PictaStudio\VenditioCore\Http\Controllers\Api\Controller;
-use PictaStudio\VenditioCore\Http\Requests\V1\Address\StoreAddressRequest;
+use PictaStudio\VenditioCore\Http\Requests\V1\Brand\StoreBrandRequest;
+use PictaStudio\VenditioCore\Http\Requests\V1\Brand\UpdateBrandRequest;
 use PictaStudio\VenditioCore\Http\Resources\V1\BrandResource;
 use PictaStudio\VenditioCore\Packages\Simple\Models\Brand;
 
+use function PictaStudio\VenditioCore\Helpers\Functions\resolve_model;
 use function PictaStudio\VenditioCore\Helpers\Functions\query;
 
 class BrandController extends Controller
 {
     public function index(): JsonResource|JsonResponse
     {
-        $filters = request()->all();
+        $this->authorize('viewAny', resolve_model('brand'));
 
-        // $this->validateData($filters, [
-        //     'all' => [
-        //         'boolean',
-        //     ],
-        //     'id' => [
-        //         'array',
-        //     ],
-        //     'id.*' => [
-        //         Rule::exists('brands', 'id'),
-        //     ],
-        // ]);
+        $filters = request()->all();
 
         return BrandResource::collection(
             $this->applyBaseFilters(query('brand'), $filters, 'brand')
         );
     }
 
-    // public function store(StoreAddressRequest $request): JsonResource
-    // {
-    //     return BrandResource::make();
-    // }
+    public function store(StoreBrandRequest $request): JsonResource
+    {
+        $this->authorize('create', resolve_model('brand'));
+
+        return BrandResource::make(
+            BrandDto::fromArray($request->validated())->create()
+        );
+    }
 
     public function show(Brand $brand): JsonResource
     {
+        $this->authorize('view', $brand);
+
         return BrandResource::make($brand);
     }
 
-    // public function update(UpdateBrandRequest $request, Brand $brand): JsonResource
-    // {
-    //     return BrandResource::make();
-    // }
+    public function update(UpdateBrandRequest $request, Brand $brand): JsonResource
+    {
+        $this->authorize('update', $brand);
+
+        return BrandResource::make(
+            BrandDto::fromArray(array_merge(
+                $request->validated(), ['brand' => $brand]
+            ))->update()
+        );
+    }
 
     public function destroy(Brand $brand)
     {
+        $this->authorize('delete', $brand);
+
         $brand->delete();
 
         return response()->noContent();
